Move Addresser/Storage to traits.rs
